Ship type declarations and stop the demo build deleting them
Two packaging bugs left `dist/` in a state that didn't match what
`package.json` promises.

`src/@types/global.d.ts` declared the `Window` properties the README
tells consumers to set, but `tsc` does not copy input `.d.ts` files to
`outDir`, and nothing in the emitted output referenced it. Consumers
following the documented configuration example got `TS2339: Property
'ANCHOR_POSITIONING_POLYFILL_OPTIONS' does not exist on type 'Window'`.
Move it to a compiled module and pull it into each entry with a
side-effect import, which `tsc` preserves in the declaration output, so
`dist/index.d.ts` and `dist/index-fn.d.ts` now reference `./global.js`.

Only the library builds set `emptyOutDir: false`, so `build:demo` — the
first step of `build` — wiped any declarations a previous `npm run
types` had emitted. `prepack` was correct only by ordering luck. Make
every build additive and empty `dist/` once, up front, via a `clean`
step.

Verified against a consumer installed from `npm pack`: it typechecks the
documented `window.ANCHOR_POSITIONING_POLYFILL_OPTIONS` assignment and
rejects an unknown option, and fails as before when the side-effect
import is removed.
