+++ package.json
@@ -49,7 +49,8 @@
     "package.json"
   ],
   "scripts": {
-    "build": "run-s build:demo build:dist build:fn",
+    "clean": "node -e \"require('node:fs').rmSync('dist', { recursive: true, force: true })\"",
+    "build": "run-s clean build:demo build:dist build:fn",
     "build:dist": "vite build",
     "build:fn": "cross-env BUILD_FN=1 vite build",
     "build:wpt": "cross-env BUILD_WPT=1 vite build",
