+++ src/index.ts
@@ -1,3 +1,5 @@
+import './global.js';
+
 import { polyfill } from './polyfill.js';
 
 // apply polyfill
