+++ tests/tsconfig.json
@@ -13,6 +13,7 @@
   "include": [
     "./**/*.ts",
     "./../src/@types/",
+    "./../src/global.ts",
     "./../*.config.ts",
     "./../*.config.js"
   ]
